Product repository

+++ pkg/data/user/user.go
@@ -2,7 +2,7 @@ package user
 
 // User is a model of a user of the service.
 type User struct {
-	ID       int    `gorm:"column:id,primaryKey" json:"id"`
+	ID       int64  `gorm:"column:id,primaryKey" json:"id"`
 	Login    string `gorm:"column:login" json:"login"`
 	Password string `gorm:"column:password" json:"password"`
 	Name     string `gorm:"column:name" json:"name"`
